git-svn-id: https://flush.svn.sourceforge.net/svnroot/flush/trunk@10 21f5d154-ea94-4097-a85b-94d12d4401c9

+++ src/mlib/types.hpp
@@ -96,7 +96,7 @@
 
 	namespace m
 	{
-		typedef int64_t		Size;
+		typedef long long	Size;
 		typedef long double	Size_float;
 		typedef int32_t		Speed;
